Issue #8746 by turegjorup: Fixed Doctrine mapping for Media entity. Corrected Annotations to XML. Added "in use" check to delete for media

+++ src/Indholdskanalen/MainBundle/Controller/MediaController.php
@@ -197,14 +197,13 @@ class MediaController extends Controller {
    */
   public function MediaDeleteAction($id) {
     $em = $this->getDoctrine()->getManager();
-    $media = $this->getDoctrine()->getRepository('ApplicationSonataMediaBundle:Media')
-      ->findOneById($id);
+    $media = $this->getDoctrine()->getRepository('ApplicationSonataMediaBundle:Media')->findOneById($id);
 
     // Create response.
     $response = new Response();
     $response->headers->set('Content-Type', 'application/json');
 
-    if ($media) {
+    if ($media && $media->getMediaOrders()->isEmpty()) {
       $em->remove(($media));
       $em->flush();
       $response->setContent(json_encode(array()));
